+++ simple_plugins/models.py
@@ -18,6 +18,10 @@ def get_page_teaser_choices():
     return _get_choices('SimplePageTeaser')
 
 
+def get_url_teaser_choices():
+    return _get_choices('SimpleURLTeaser')
+
+
 def get_text_and_image_choices():
     return _get_choices('SimpleTextAndImage')
 
@@ -48,6 +52,26 @@ class SimplePageTeaserPluginModel(CMSPlugin):
     @property
     def title(self):
         return self.override_title or self.page.title_set.first().title
+
+
+class SimpleURLTeaserPluginModel(CMSPlugin):
+    flavor = models.PositiveSmallIntegerField(blank=False)
+    url = models.URLField(blank=False)
+    title = models.CharField(max_length=80, blank=True)
+    subtitle = models.CharField(max_length=80, blank=True)
+    content = HTMLField(blank=False)
+    image = FilerImageField(null=False, blank=False)
+
+    @staticmethod
+    def get_flavor_choices_fun():
+        return lazy(get_url_teaser_choices, list)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._meta.get_field('flavor')._choices = self.get_flavor_choices_fun()()
+
+    def __str__(self):
+        return self.title
 
 
 class TextAndImagePluginModel(CMSPlugin):